For scala 2 version, moved DiagrammedExpr.scala and DiagrammedExprMacro.scala into org.scalatest.diagrams package, and deprecated the ones in org.scalatest package.
